voice-analyzer/app/requirements.txt
fastapi
uvicorn[standard]
librosa
numpy
scipy
soundfile
praat-parselmouth
python-multipart